requirements.txt
Django==2.2.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
django-cors-headers==3.1.0
djoser==2.0.3
